feat: 지도 롱클릭 시 번개 모각코 생성 위치 표시
